// src/app/pages/approval/models/vacaciones/response-vacacion-solicitada.interface.ts
export interface IVacacionSolicitada {
  idPeriodo: string;
  idPersona: string;
  fechaInicio: string;
  fechaFin: string;
  aprobadoPorJefe: string;
  aprobadoPorRRHH: string;
  registroPersona: string;
  registroFecha: string;
  aprobacionJefeFecha: string;
  aprobacionRRHHFecha: string;
  estado: string;
}